rename main_window module
